move feature enable test to separate directory
