+++ Cargo.toml
@@ -85,6 +85,7 @@ directories = "5.0.0"
 dirs = "5.0.0"
 duct = { version = "0.13", default-features = false }
 event-listener = "2.5.1"
+eyre = { version = "0.6", default-features = false}
 flate2 = { version = "1.0.25" }
 futures = { version = "0.3.1" }
 futures-util = { version = "0.3.6", default-features = false }
